Fix Alpine ARM64 crossbuild by updating binutils
The linker from binutils 2.27 that we were using before was producing
binaries with missing symbols and so it was not possible to load
libcoreclr.so.

We were also not building the binutils with plugins enabled, which
caused LTO to be disabled.

This change updates binutils to 2.31.1 and adds missing option to the
configure command line so that the plugins support is built in.