Unify clang in containers to version 9
For CentOS, we need to build clang 9 and lldb 9 from sources. On CentOS 6,
it is more tricky, as we need to build a recent version of swig to enable
building clang 9.
For Ubuntu, we can install it from packages, but there is another issue.
The lldb-9 package has broken dependencies and cannot be installed.
So we install lldb-6.0 instead, which is sufficient for the diagnostics
repo build.