Change ubuntu cross arm alpine to use 3.13 rootfs
Due to the breaking change on Alpine 3.13 w.r.t. time_t type
size, we will only support ARM on Alpine >= 3.13.
This change updates the ROOTFS that the cross build uses
to be the one of Alpine 3.13.

I have also created cross image for Ubuntu 16.04, as we should
be cross building on 16.04 rather than 18.04 to get better coverage
of Linux distros where cross-tools like crossgen can be run.
Ideally, we would use CentOS 7, but we cannot run crossbuild on
that one.