Take the timeout up to five hours
This whole adventure has highlighted a fragility in how we deal with changes lower down the stack which impact many many images requiring many many rebuilds. Obviously on a normal day, a five hour timeout is way too long and indicates a problem, but occasionally it's needed. Is there a better way to codify that?