Add rh-ruby binaries to path on centos-7-rpmpkg
For the rpmpkg container, we need to be able to find fpm, which is now being installed within scl, and no longer in the standard place. This change adds the rh-ruby bin paths to the PATH, LD_LIBRARY_PATH, and a few others so that runtime can generate rpms on in this container.