Add libatomic to Centos Stream 8 image
The CentOS 8 image is used to run Mono LLVM AOT tests, in the interim, until Mono's AOT compiler is updated to use `llvm-as` instead of GNU Binutils `as`. However, LLVM 16 introduces a dependency on libatomic.so.1 which is satisfied on Mariner, but is absent on our CentOS Stream 8 image.

Fixing Mono's aot-compiler.c is the better long-term solution, but this unblocks test failures on the LLVM 16 integration work.