Add images with both "host" and "target" crossroots in one image.
This is required in some cases - e.g. compiling node.js for ARM64/Alpine,
and the ever-broken ARM64 runtime perf lane (where we need headers for
ARM64 available alongside linkable libraries for AMD64, depending on
the build phase).

As is, we've achieved the same result largely by accident in the
pre-Mariner build images, where we have enough x64 in the host OS to link
successfully, and the ARM64 files in the crossroot.

The precedent on this is https://github.com/dotnet/dotnet-buildtools-prereqs-docker/pull/901
which fixed our Android cross-compiler builds, by offering both Android
and $HOST in the same system

If at some point in the future we move from building on x64 to building
on ARM64, we will need the equivalent images for that.