Replace alpine tags in Azure Linux images with musl (#1270)
We want to move to `musl` tags since that naming better reflects
the libc compatibility provided by these images. This replaces
`alpine` with `musl` in the .NET 10 build images, and adds `musl`
variants (keeping the old `alpine` tags for compat) to the .NET 8
and .NET 9 images. This doesn't change the cbl-mariner-2.0 or
ubuntu cross images because we're planning to remove those.